working on rendering drawing from json file

+++ src/model/color.rs
@@ -27,42 +27,42 @@ impl Color {
     pub fn mutate(&mut self) -> bool {
         let mut mutation_happened = false;
 
-        if rand::thread_rng().gen::<f64>() < CHANGE_COLOR_PROB {
+        if rand::thread_rng().gen::<f32>() < CHANGE_COLOR_PROB {
             self.r = rand::thread_rng().gen::<u8>();
             mutation_happened = true;
         }
-        if rand::thread_rng().gen::<f64>() < CHANGE_COLOR_PROB {
+        if rand::thread_rng().gen::<f32>() < CHANGE_COLOR_PROB {
             self.g = rand::thread_rng().gen::<u8>();
             mutation_happened = true;
         }
-        if rand::thread_rng().gen::<f64>() < CHANGE_COLOR_PROB {
+        if rand::thread_rng().gen::<f32>() < CHANGE_COLOR_PROB {
             self.b = rand::thread_rng().gen::<u8>();
             mutation_happened = true;
         }
-        if rand::thread_rng().gen::<f64>() < CHANGE_COLOR_PROB {
+        if rand::thread_rng().gen::<f32>() < CHANGE_COLOR_PROB {
             self.a = rand::thread_rng().gen::<u8>().clamp(MIN_ALPHA, MAX_ALPHA);
             mutation_happened = true;
         }
 
         //// same but micro adjustments
-        if rand::thread_rng().gen::<f64>() < MICRO_ADJUSTMENT_PROBABILITY {
+        if rand::thread_rng().gen::<f32>() < MICRO_ADJUSTMENT_PROBABILITY {
             self.r = Color::micro_adjust(self.r);
             mutation_happened = true;
         }
-        if rand::thread_rng().gen::<f64>() < MICRO_ADJUSTMENT_PROBABILITY {
+        if rand::thread_rng().gen::<f32>() < MICRO_ADJUSTMENT_PROBABILITY {
             self.g = Color::micro_adjust(self.g);
             mutation_happened = true;
         }
-        if rand::thread_rng().gen::<f64>() < MICRO_ADJUSTMENT_PROBABILITY {
+        if rand::thread_rng().gen::<f32>() < MICRO_ADJUSTMENT_PROBABILITY {
             self.b = Color::micro_adjust(self.b);
             mutation_happened = true;
         }
-        if rand::thread_rng().gen::<f64>() < MICRO_ADJUSTMENT_PROBABILITY {
+        if rand::thread_rng().gen::<f32>() < MICRO_ADJUSTMENT_PROBABILITY {
             self.a = Color::micro_adjust(self.a).clamp(MIN_ALPHA, MAX_ALPHA);
             mutation_happened = true;
         }
         ////
-        if rand::thread_rng().gen::<f64>() < LIGHTEN_COLOR_PROB {
+        if rand::thread_rng().gen::<f32>() < LIGHTEN_COLOR_PROB {
             if self.r < u8::MAX && self.g < u8::MAX && self.b < u8::MAX {
                 self.r += 1;
                 self.g += 1;
@@ -70,7 +70,7 @@ impl Color {
                 mutation_happened = true;
             }
         }
-        if rand::thread_rng().gen::<f64>() < DARKEN_COLOR_PROB {
+        if rand::thread_rng().gen::<f32>() < DARKEN_COLOR_PROB {
             if self.r > u8::MIN && self.g > u8::MIN && self.b > u8::MIN {
                 self.r -= 1;
                 self.g -= 1;
@@ -84,7 +84,7 @@ impl Color {
 
     // increment or decrement with 50% chance while avoiding overflows and underflows
     fn micro_adjust(mut val: u8) -> u8 {
-        val = if rand::thread_rng().gen::<f64>() > 0.5 {
+        val = if rand::thread_rng().gen::<f32>() > 0.5 {
             if val < u8::MAX {
                 val + 1
             } else {
